Change desktop size when switching from a disconnected display
When an active display is disconnected, xrandr somehow fails to
resize the desktop (probably thinking it hasn't got to do so, as
there is only one display). If the primary screen is smaller than
the disconnected one, this results in part of the display not
being visible.

I have found no better way than the damned xrandr -s $RES, that I
wanted to avoid at all costs for not knowing the effect of this on
monitor configuration.